+++ src/vmware/azext_vmware/aaz/latest/vmware/addon/_create.py
@@ -12,13 +12,13 @@ from azure.cli.core.aaz import *
 
 
 class Create(AAZCommand):
-    """Create a addon in a private cloud
+    """Create an addon in a private cloud
     """
 
     _aaz_info = {
-        "version": "2023-03-01",
+        "version": "2023-09-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/addons/{}", "2023-03-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.avs/privateclouds/{}/addons/{}", "2023-09-01"],
         ]
     }
 
@@ -51,6 +51,9 @@ class Create(AAZCommand):
             options=["-c", "--private-cloud"],
             help="Name of the private cloud",
             required=True,
+            fmt=AAZStrArgFormat(
+                pattern="^[-\w\._]+$",
+            ),
         )
         _args_schema.resource_group = AAZResourceGroupNameArg(
             required=True,
@@ -77,14 +80,13 @@ class Create(AAZCommand):
         _args_schema.vr = AAZObjectArg(
             options=["--vr"],
             arg_group="Properties",
-            help="a vSphere Replication (VR) addon for a private cloud",
+            help="a vSphere Replication (VR) addon for a private cloud.",
         )
 
         arc = cls._args_schema.arc
         arc.vcenter = AAZStrArg(
             options=["vcenter"],
             help="The VMware vCenter resource ID",
-            required=True,
         )
 
         hcx = cls._args_schema.hcx
@@ -98,7 +100,6 @@ class Create(AAZCommand):
         srm.license_key = AAZStrArg(
             options=["license-key"],
             help="The Site Recovery Manager (SRM) license",
-            required=True,
         )
 
         vr = cls._args_schema.vr
@@ -194,7 +195,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-03-01",
+                    "api-version", "2023-09-01",
                     required=True,
                 ),
             }
@@ -234,7 +235,7 @@ class Create(AAZCommand):
 
             disc_arc = _builder.get(".properties{addonType:Arc}")
             if disc_arc is not None:
-                disc_arc.set_prop("vCenter", AAZStrType, ".arc.vcenter", typ_kwargs={"flags": {"required": True}})
+                disc_arc.set_prop("vCenter", AAZStrType, ".arc.vcenter")
 
             disc_hcx = _builder.get(".properties{addonType:HCX}")
             if disc_hcx is not None:
@@ -242,7 +243,7 @@ class Create(AAZCommand):
 
             disc_srm = _builder.get(".properties{addonType:SRM}")
             if disc_srm is not None:
-                disc_srm.set_prop("licenseKey", AAZStrType, ".srm.license_key", typ_kwargs={"flags": {"required": True}})
+                disc_srm.set_prop("licenseKey", AAZStrType, ".srm.license_key")
 
             disc_vr = _builder.get(".properties{addonType:VR}")
             if disc_vr is not None:
@@ -275,6 +276,10 @@ class Create(AAZCommand):
                 flags={"read_only": True},
             )
             _schema_on_200_201.properties = AAZObjectType()
+            _schema_on_200_201.system_data = AAZObjectType(
+                serialized_name="systemData",
+                flags={"read_only": True},
+            )
             _schema_on_200_201.type = AAZStrType(
                 flags={"read_only": True},
             )
@@ -292,7 +297,6 @@ class Create(AAZCommand):
             disc_arc = cls._schema_on_200_201.properties.discriminate_by("addon_type", "Arc")
             disc_arc.v_center = AAZStrType(
                 serialized_name="vCenter",
-                flags={"required": True},
             )
 
             disc_hcx = cls._schema_on_200_201.properties.discriminate_by("addon_type", "HCX")
@@ -303,13 +307,32 @@ class Create(AAZCommand):
             disc_srm = cls._schema_on_200_201.properties.discriminate_by("addon_type", "SRM")
             disc_srm.license_key = AAZStrType(
                 serialized_name="licenseKey",
-                flags={"required": True},
             )
 
             disc_vr = cls._schema_on_200_201.properties.discriminate_by("addon_type", "VR")
             disc_vr.vrs_count = AAZIntType(
                 serialized_name="vrsCount",
                 flags={"required": True},
+            )
+
+            system_data = cls._schema_on_200_201.system_data
+            system_data.created_at = AAZStrType(
+                serialized_name="createdAt",
+            )
+            system_data.created_by = AAZStrType(
+                serialized_name="createdBy",
+            )
+            system_data.created_by_type = AAZStrType(
+                serialized_name="createdByType",
+            )
+            system_data.last_modified_at = AAZStrType(
+                serialized_name="lastModifiedAt",
+            )
+            system_data.last_modified_by = AAZStrType(
+                serialized_name="lastModifiedBy",
+            )
+            system_data.last_modified_by_type = AAZStrType(
+                serialized_name="lastModifiedByType",
             )
 
             return cls._schema_on_200_201